add session scoped auth to all routes
