+++ src/app/api/user_locations/route.ts
@@ -1,25 +1,38 @@
 import { UserLocation } from '@/app/Classes/UserLocation';
-import { getUserLocationById } from '@/app/Util/DatabaseApiCalls';
+import { authOptions } from '@/app/lib/authOptions';
 import { sql, db } from '@vercel/postgres';
+import { getServerSession } from 'next-auth/next';
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function GET(request: NextRequest) {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+  }
+
+  const userId = request.nextUrl.searchParams.get('user_id');
+  if (!userId) {
+    return NextResponse.json({ error: 'Missing user_id parameter' }, { status: 400 });
+  }
+
+  if (session.user.id !== userId) {
+    return NextResponse.json({ error: 'Forbidden' }, { status: 403 });
+  }
+
   try {
-    const userId = request.nextUrl.searchParams.get('user_id');
     const id = request.nextUrl.searchParams.get('id');
-    let matchingLocs;
     let foundEntries;
-    if (userId && id) {
-      matchingLocs =
+    if (id) {
+      const matchingLocs =
         await sql`SELECT * FROM sendtemps.user_locations WHERE user_id = ${userId} AND id = ${id};`;
       foundEntries = matchingLocs?.rows[0];
-    } else if (userId && !id) {
-      matchingLocs = await sql`SELECT * FROM sendtemps.user_locations WHERE user_id = ${userId};`;
+    } else {
+      const matchingLocs =
+        await sql`SELECT * FROM sendtemps.user_locations WHERE user_id = ${userId};`;
       foundEntries = matchingLocs?.rows;
     }
 
-    const response = NextResponse.json(foundEntries, { status: 200 });
-    return response;
+    return NextResponse.json(foundEntries, { status: 200 });
   } catch (error) {
     console.error(error);
     return NextResponse.json({ error }, { status: 500 });
@@ -27,6 +40,11 @@ export async function GET(request: NextRequest) {
 }
 
 export async function POST(request: NextRequest) {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+  }
+
   try {
     const reqBody = await request.json();
     const newUserLoc = new UserLocation(
@@ -34,20 +52,16 @@ export async function POST(request: NextRequest) {
       reqBody.name,
       reqBody.latitude,
       reqBody.longitude,
-      reqBody.user_id,
+      session.user.id,
       reqBody.poi_type,
       null,
       null,
     );
-    let response;
-    if (newUserLoc) {
-      await sql`INSERT INTO sendtemps.user_locations (name, latitude, longitude, user_id, poi_type, date_created, last_modified) VALUES (${newUserLoc.name}, ${newUserLoc.latitude}, ${newUserLoc.longitude}, ${newUserLoc.user_id}, ${newUserLoc.poi_type}, ${newUserLoc.date_created}, ${newUserLoc.last_modified})`;
-      response = NextResponse.json(
-        `Success - New Location "${newUserLoc.name}" created for user: ${newUserLoc.user_id}`,
-        { status: 201 },
-      );
-    }
-    return response;
+    await sql`INSERT INTO sendtemps.user_locations (name, latitude, longitude, user_id, poi_type, date_created, last_modified) VALUES (${newUserLoc.name}, ${newUserLoc.latitude}, ${newUserLoc.longitude}, ${newUserLoc.user_id}, ${newUserLoc.poi_type}, ${newUserLoc.date_created}, ${newUserLoc.last_modified})`;
+    return NextResponse.json(
+      `Success - New Location "${newUserLoc.name}" created for user: ${newUserLoc.user_id}`,
+      { status: 201 },
+    );
   } catch (error) {
     console.error(error);
     return NextResponse.json({ error }, { status: 500 });
@@ -55,50 +69,69 @@ export async function POST(request: NextRequest) {
 }
 
 export async function PATCH(request: NextRequest) {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+  }
+
   try {
     const reqBody = await request.json();
     const validCols = ['name', 'poi_type'];
-    let userLoc = await getUserLocationById(reqBody.userId, reqBody.id);
 
-    if (userLoc && validCols.includes(reqBody.changeCol)) {
-      let patchLoc = new UserLocation(
-        userLoc.id,
-        userLoc.name,
-        userLoc.latitude,
-        userLoc.longitude,
-        userLoc.user_id,
-        userLoc.poi_type,
-        userLoc.date_created,
-        userLoc.last_modified,
-      );
-      if (reqBody.changeCol === 'name') {
-        patchLoc.updateName(reqBody.data);
-      } else if (reqBody.changeCol === 'poi_type') {
-        patchLoc.updatePOIType(reqBody.data);
-      }
-      patchLoc.updateLastModified();
+    if (!reqBody.id || !reqBody.changeCol || !validCols.includes(reqBody.changeCol)) {
+      return NextResponse.json({ error: 'Invalid request body' }, { status: 400 });
+    }
 
-      // Connect to DB and start SQl queries
-      const client = await db.connect();
+    const { rows } = await sql`
+      SELECT * FROM sendtemps.user_locations WHERE id = ${reqBody.id};
+    `;
+    const userLoc = rows[0] ?? null;
 
-      if (reqBody.changeCol === 'name') {
-        await client.sql`UPDATE sendtemps.user_locations 
+    if (!userLoc) {
+      return NextResponse.json({ error: 'Location not found' }, { status: 404 });
+    }
+
+    if (userLoc.user_id !== session.user.id) {
+      return NextResponse.json({ error: 'Forbidden' }, { status: 403 });
+    }
+
+    const patchLoc = new UserLocation(
+      userLoc.id,
+      userLoc.name,
+      userLoc.latitude,
+      userLoc.longitude,
+      userLoc.user_id,
+      userLoc.poi_type,
+      userLoc.date_created,
+      userLoc.last_modified,
+    );
+
+    if (reqBody.changeCol === 'name') {
+      patchLoc.updateName(reqBody.data);
+    } else if (reqBody.changeCol === 'poi_type') {
+      patchLoc.updatePOIType(reqBody.data);
+    }
+    patchLoc.updateLastModified();
+
+    const client = await db.connect();
+
+    if (reqBody.changeCol === 'name') {
+      await client.sql`UPDATE sendtemps.user_locations 
         SET name = ${reqBody.data} 
         WHERE id = ${patchLoc.id} AND user_id = ${patchLoc.user_id};`;
-      } else if (reqBody.changeCol === 'poi_type') {
-        await client.sql`UPDATE sendtemps.user_locations 
+    } else if (reqBody.changeCol === 'poi_type') {
+      await client.sql`UPDATE sendtemps.user_locations 
         SET poi_type = ${reqBody.data} 
         WHERE id = ${patchLoc.id} AND user_id = ${patchLoc.user_id};`;
-      }
-
-      await client.sql`UPDATE sendtemps.user_locations 
-        SET last_modified = ${patchLoc.last_modified} 
-        WHERE id = ${patchLoc.id} AND user_id = ${patchLoc.user_id};`;
-
-      client.release();
-
-      return NextResponse.json({ patchLoc }, { status: 200 });
     }
+
+    await client.sql`UPDATE sendtemps.user_locations 
+      SET last_modified = ${patchLoc.last_modified} 
+      WHERE id = ${patchLoc.id} AND user_id = ${patchLoc.user_id};`;
+
+    client.release();
+
+    return NextResponse.json({ patchLoc }, { status: 200 });
   } catch (error) {
     console.error(error);
     return NextResponse.json({ error }, { status: 500 });
@@ -106,8 +139,18 @@ export async function PATCH(request: NextRequest) {
 }
 
 export async function DELETE(request: NextRequest) {
+  const session = await getServerSession(authOptions);
+  if (!session) {
+    return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
+  }
+
   try {
     const deleteLoc = await request.json();
+
+    if (deleteLoc.user_id !== session.user.id) {
+      return NextResponse.json({ error: 'Forbidden' }, { status: 403 });
+    }
+
     await sql`DELETE FROM sendtemps.user_locations 
       WHERE id = ${deleteLoc.id} AND user_id = ${deleteLoc.user_id};`;
     return NextResponse.json(
